pyproject.toml
[project]
name = "unity-vault-bot"
version = "1.0.0"
description = "Production Discord bot for Unity Vault community"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "discord.py[voice]>=2.3.0",
    "asyncpg>=0.29.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]

[tool.ruff]
target-version = "py312"
line-length = 100
